Changed the way CLOSURE_APPLYn works.
CLOSURE_APPLYn when applied to a function requiring less than n arguments
will now call DO_EVAL which will do the right thing in that case.
This seems to work.

Added BIGINT to DO_EVAL as an already evaluated value.

Changed the code generator so names for imports are not doubly encoded.