Use System.FilePath.replaceExtension in Main
Previously, the `epic` executable had its own logic to strip an
extension, by splitting on the `.` character. However, this logic would
split on the *first* `.`, when it should split on the *last* `.`. This
leads to problems, such as:

    $ epic /nix/store/r7dwf24y88bwdzfb1s9cwk3sm35pcnwg-haskell-Agda-ghc7.8.3-2.4.2-shared/share/x86_64-linux-ghc-7.8.3/Agda-2.4.2/EpicInclude/AgdaPrelude.e
    Fatal error: can't create /nix/store/r7dwf24y88bwdzfb1s9cwk3sm35pcnwg-haskell-Agda-ghc7.o: Read-only file system

Which is not the output path that we would expect.

Rather than reinventing the wheel and implementing file path handling
ourselves, we now depend on `filepath` for this logic. `filepath` is
already a transitive dependency, so the dependency graph is unchanged
with this change.